refactor: 패키지 위치 및 포맷팅 수정
