+++ src/main/java/web/WebApplication.java
@@ -43,8 +43,6 @@ public class WebApplication {
         }, jsonTransformer);
     }
 
-
-
     private static String render(Map<String, Object> model, String templatePath) {
         return new HandlebarsTemplateEngine().render(new ModelAndView(model, templatePath));
     }
