+++ src/test/java/web/DBConnectionSetUpTest.java
@@ -1,4 +1,4 @@
-package chess;
+package web;
 
 import static org.assertj.core.api.Assertions.assertThat;
